ID-1: All pages moved to folder
