--- backup_package_json_backup
@@ -1,49 +1,0 @@
-{
-  "name": "iqbal-demystified-react",
-  "version": "0.1.0",
-  "private": true,
-  "proxy": "http://icanmakemyownapp.com/iqbal/v3",
-  "dependencies": {
-    "@material-ui/core": "^4.1.1",
-    "bootstrap": "^4.3.1",
-    "history": "^4.9.0",
-    "jquery": "^3.4.1",
-    "react": "^16.8.6",
-    "react-bootstrap": "^1.0.0-beta.8",
-    "react-dom": "^16.8.6",
-    "react-router-dom": "^5.0.0",
-    "react-scripts": "3.0.1",
-    "react-table": "^6.10.0",
-    "yaml": "^1.6.0",
-    "yamljs": "^0.3.0"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "devDependencies": {
-    "raw-loader": "^2.0.0"
-  },
-  "rnpm": {
-    "assets": [
-      "assets/fonts/Sarpanch"
-    ]
-  }
-}
