--- src/TabView1Old.css
@@ -1,59 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.App-logo {
-  /*animation: App-logo-spin infinite 20s linear;*/
-  padding: 10rem 10rem;
-  height: 40vmin;
-  /*pointer-events: none;*/
-  pointer-events: all;
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-.tab-list {
-	  border-bottom: 1px solid #ccc;
-	  color: black;
-	    padding-left: 10;
-	      font-size: 100px;
-    }
-
-    .tab-list-item { display: inline-block;
-	        list-style: none;
-		  color: black;
-		  margin-bottom: -1px;
-		    padding: 10rem 10rem;
-		    font-size: 100px;
-	    }
-
-	    .tab-list-active {
-		      background-color: white;
-		        color: red;
-		        border: solid #ccc;
-			  border-width: 10px 10px 1 10px;
-		    	  padding: 10rem 10rem;
-			  font-size: 100px;
-		  }
