--- src/contact.js
@@ -1,11 +1,0 @@
-import React from 'react'
-export default class Contact extends React.Component {
-	  render() {
-		      return ( 	
-			  	<h1>
-			  		Hello this is my Contact
-			  	</h1>
-		      )
-			    }
-}
-// export default Contact
