+++ src/misc/CommentPage.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import StaticContentService from '../StaticContentServiceYaml'
+import StaticContentService from './StaticContentServiceYaml'
 // import Tabs from './Tabs'
 
 // for formatting
