+++ src/poetry/PoemPage.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import StaticContentService from '../StaticContentServiceYaml'
+import StaticContentService from '../misc/StaticContentServiceYaml'
 
 // for formatting
 import '../main_page/TopSectionMainPage/TopSectionMainPage.css';
